chore(deps)!: upgrade dependant actions to new major versions
This upgrades `tibdex/github-app-token` to `v2` and
`google-github-actions/release-please-action` to `v4`.

Among other things, this ensures that we are compatible with the GitHub
Actions upgrade to Node 20, and avoiding issues when Node 16-based
actions are deprecated sometime this spring.

BREAKING CHANGE: Underlying actions have been upgraded, requiring changes to available inputs