ci(release): switch to `main` version of self
Since Google moved their upstream action, prior versions of
`jimeh/release-please-manifest-action` will not work until a new one is
cut.

But since this action uses itself to release itself, we have to force it
to use the action from the main branch to be able to cut a new release.